moving GapfillMethod from time series crate to extension asap code
