+++ extension/src/asap.rs
@@ -7,7 +7,7 @@ use crate::{
     aggregate_utils::in_aggregate_context, palloc::{Internal, InternalAsValue, Inner, ToInternal},
 };
 
-use time_series::{TSPoint, GapfillMethod};
+use time_series::TSPoint;
 
 use crate::time_series::{Timevector, TimevectorData, SeriesType};
 
@@ -23,6 +23,83 @@ pub fn asap_smooth_raw(
 // hack to allow us to qualify names with "toolkit_experimental"
 // so that pgx generates the correct SQL
 mod toolkit_experimental {
+}
+
+pub enum GapfillMethod {
+    LOCF,
+    Linear,
+    Nearest,  // Defaults to left on tie
+}
+
+impl GapfillMethod {
+    // Adds the given number of points to the end of a non-empty NormalTimevector
+    pub fn fill_normalized_series_gap(&self, values: &mut Vec<f64>, points: i32, post_gap_val: f64) {
+        assert!(!values.is_empty());
+        let last_val = *values.last().unwrap();
+        for i in 1..=points {
+            match self {
+                GapfillMethod::LOCF => values.push(last_val),
+                GapfillMethod::Linear => values.push(last_val + (post_gap_val - last_val) * i as f64 / (points + 1) as f64),
+                GapfillMethod::Nearest => values.push(if i <= (points + 1) / 2 {last_val} else {post_gap_val}),
+            }
+        }
+    }
+}
+
+impl GapfillMethod {
+    // Determine a value to the left of a given point or two (for linear) using the given gapfill method
+    // TODO: this returns the first value for LOCF, which probabaly isn't correct, technically this function shouldn't be valid for LOCF,
+    pub fn predict_left(&self, target_time: i64, first: TSPoint, second: Option<TSPoint>) -> TSPoint {
+        TSPoint {
+            ts: target_time,
+            val: match self {
+                GapfillMethod::LOCF => first.val,
+                GapfillMethod::Nearest => first.val,
+                GapfillMethod::Linear => {
+                    let second = match second {
+                        Some(v) => v,
+                        None => panic!{"Unable to predict left point without two values to interpolate from"},
+                    };
+                    let slope = (first.val - second.val) / (first.ts - second.ts) as f64;
+                    first.val - slope * (first.ts - target_time) as f64
+                },
+            }
+        }
+    }
+
+    // Determine a value to the right of a given point or two (for linear) using the given gapfill method
+    pub fn predict_right(&self, target_time: i64, last: TSPoint, penultimate: Option<TSPoint>) -> TSPoint {
+        TSPoint {
+            ts: target_time,
+            val: match self {
+                GapfillMethod::LOCF => last.val,
+                GapfillMethod::Nearest => last.val,
+                GapfillMethod::Linear => {
+                    let penultimate = match penultimate {
+                        Some(v) => v,
+                        None => panic!{"Unable to predict right point without two values to interpolate from"},
+                    };
+                    let slope = (last.val - penultimate.val) / (last.ts - penultimate.ts) as f64;
+                    last.val + slope * (target_time - last.ts) as f64
+                },
+            }
+        }
+    }
+
+    // Given a target time and the immediate points to either side, provide the missing point
+    pub fn gapfill(&self, target_time: i64, left: TSPoint, right: TSPoint) -> TSPoint {
+        TSPoint {
+            ts: target_time,
+            val: match self {
+                GapfillMethod::LOCF => left.val,
+                GapfillMethod::Nearest => if target_time - left.ts <= right.ts - target_time {left.val} else {right.val},
+                GapfillMethod::Linear => {
+                    let slope = (right.val - left.val) / (right.ts - left.ts) as f64;
+                    left.val + slope * (target_time - left.ts) as f64
+                },
+            }
+        }
+    }
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
